Split monolithic tests into separate files
